requirements.txt
Flask==1.1.1
pyserial==3.4
